requirements.txt
PyQt6>=6.4.0
mysql-connector-python>=8.0.33
Pillow>=9.5.0
qtawesome>=1.2.0
matplotlib>=3.7.0
